+++ src/common/config_parser.cpp
@@ -1,4 +1,5 @@
 #include "common/authors.h"
+#include "common/build_info.h"
 #include "common/cli_helper.h"
 #include "common/config.h"
 #include "common/config_parser.h"
@@ -111,6 +112,9 @@ void ConfigParser::addOptionsGeneral(cli::CLIWrapper& cli) {
     "Print list of authors and exit");
   cli.add<bool>("--cite",
     "Print citation and exit");
+  cli.add<std::string>("--build-info",
+    "Print CMake build options and exit. Set to 'all' to print advanced options")
+    ->implicit_val("basic");
   cli.add<std::vector<std::string>>("--config,-c",
     "Configuration file(s). If multiple, later overrides earlier");
   cli.add<size_t>("--workspace,-w",
@@ -504,7 +508,7 @@ void ConfigParser::addOptionsTraining(cli::CLIWrapper& cli) {
      true);
 
   // add ULR settings
-  addSuboptionsULR(cli); 
+  addSuboptionsULR(cli);
 
   cli.add<std::vector<std::string>>("--task",
      "Use predefined set of options. Possible values: transformer, transformer-big");
@@ -832,6 +836,15 @@ Ptr<Options> ConfigParser::parseOptions(int argc, char** argv, bool doValidate){
 
   if(get<bool>("cite")) {
     std::cerr << citation() << std::endl;
+    exit(0);
+  }
+
+  auto buildInfo = get<std::string>("build-info");
+  if(!buildInfo.empty() && buildInfo != "false") {
+    if(buildInfo == "all")
+      std::cerr << cmakeBuildOptionsAdvanced() << std::endl;
+    else
+      std::cerr << cmakeBuildOptions() << std::endl;
     exit(0);
   }